OffsetLength is calculated in its own contributor now
