+++ server_plugin/test/completion_contributor_test_util.dart
@@ -41,7 +41,8 @@ abstract class AbstractCompletionContributorTest
   List<CompletionContributor> createContributors() => <CompletionContributor>[
         new AngularCompletionContributor(),
         new NgInheritedReferenceContributor(),
-        new NgTypeMemberContributor()
+        new NgTypeMemberContributor(),
+        new NgOffsetLengthContributor(),
       ];
 
   @override
